move settings method to utils file
